refactor components expect for charts
